requirements.txt
openai
python-dotenv
tqdm 